% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_downloading_load_data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(
  dat,
  left,
  right,
  rna_umap,
  adt_umap,
  grouping_variable = NULL,
  grouping_values = NULL,
  min_counts = NULL
)
}
\arguments{
\item{left}{annotation in the left of the diagram}

\item{right}{annotation in the right of the diagram}

\item{filename}{path to data file}

\item{data_type}{experiment used in the MultiAssayExperiment or data contained in logcounts in SingleCellExperiment}

\item{pc1_axis1}{first axis of first dimension reduction space}

\item{pc1_axis2}{second axis of first dimension reduction space}

\item{pc2_axis1}{first axis of second dimension reduction space}

\item{pc2_axis2}{second axis of second dimension reduction space}
}
\value{
The return value, if any, from executing the function.
}
\description{
A utility function for reading a processing the data as `tagtango`
}
