% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_downloading_load_data.R
\name{process_data}
\alias{process_data}
\title{Process data as the shiny application}
\usage{
process_data(
  filename,
  data_type,
  left,
  right,
  pc1_axis1 = NULL,
  pc1_axis2 = NULL,
  pc2_axis1 = NULL,
  pc2_axis2 = NULL,
  filter_variable = NULL,
  filter_values = NULL,
  grouping_variable = NULL,
  grouping_values = NULL,
  min_counts = NULL
)
}
\arguments{
\item{filename}{path to data file}

\item{data_type}{experiment used in the MultiAssayExperiment or data contained in logcounts in SingleCellExperiment}

\item{left}{annotation in the left of the diagram}

\item{right}{annotation in the right of the diagram}

\item{pc1_axis1}{first axis of first dimension reduction space}

\item{pc1_axis2}{second axis of first dimension reduction space}

\item{pc2_axis1}{first axis of second dimension reduction space}

\item{pc2_axis2}{second axis of second dimension reduction space}

\item{filter_variable}{filtering variable found in `dat`}

\item{filter_values}{values in `filter_variable` that need to be excluded}

\item{grouping_variable}{grouping variable found in `dat`}

\item{grouping_values}{values in `grouping_variable` that need to be filtered}

\item{min_counts}{minimum number of cells in a link for this to be displayed}
}
\value{
Returns a list containing two attributes: data and network. The first contains the data filtered according to the grouping and filtering values. The second, contains the data organized in a manner that is readible for the sankeyNetwork.
}
\description{
A utility function for reading a processing the data as `tagtango` does.
}
