% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rose_plot.R
\name{rose_plot}
\alias{rose_plot}
\title{rose_plot}
\usage{
rose_plot(
  norm,
  data,
  selected,
  title,
  n_petals = 10,
  valley = NULL,
  palette = "RdYlGn"
)
}
\arguments{
\item{norm}{normalized expression data with cells as rows and columns as markers/genes (i.e. function `process_data()` should provide this under the attribute `data`)}

\item{data}{data.frame with the different annotations, where each row represent a cell. The rownames should match those of `norm` (i.e. function `process_data()` should provide this under the attribute `network`).}

\item{selected}{boolean array with selected cells as True values, for all cells in `data`.}

\item{title}{main plot title. Default `NULL`.}

\item{n_petals}{if an integer value, it defines the number of petals of the rose plot, selecting those that are most "relevant". If an array with marker names, it uses those.}

\item{valley}{the value in `norm` corresponding to the valley separating positive and negative peak for CITE-seq data.}

\item{palette}{color palette, default "RdYlGn"}
}
\value{
returns a ggplot object.
}
\description{
A function to generate a rose plot as `tagtango` does.
}
