% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_scatter.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Generate a ggplot scatter plot}
\usage{
scatter_plot(
  data,
  labels,
  values = c("a", "b"),
  title = "UMAP of the RNA data",
  xlabel = "first axis",
  ylabel = "second axis"
)
}
\arguments{
\item{data}{data.frame with dimension reduction axes.}

\item{labels}{array with labels for first (and potentially second selections), for all cells in `data'.}

\item{values}{boolean array with second selection cells as TRUE values, for all cells in `data'.}

\item{title}{boolean array with second selection cells as TRUE values, for all cells in `data'.}

\item{xlabel}{the text for the x axis.}

\item{ylabel}{the text for the y axis.}
}
\value{
returns a ggplot object.
}
\description{
A function to generate a scatter plot as `tagtango' does.
}
