% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test CITE-seq data}
\format{
## `test_data`
A SingleCellExperment with 7472 cells, 17 protein markers, and 33538 gene IDs:
\describe{
  \item{date}{November 19, 2018}
}
}
\source{
<https://support.10xgenomics.com/single-cell-gene-expression/datasets/3.0.0/pbmc_10k_protein_v3>
}
\usage{
test_data
}
\description{
A subset of the 10x dataset: 10k PBMCs from a healthy donor - gene expression and cell surface protein, single cell gene expression (Single Cell Gene Expression Dataset by Cell Ranger 3.0.0)
Peripheral blood mononuclear cells (PBMCs) from a healthy donor stained with TotalSeq-B antibodies.
}
\keyword{datasets}
