% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stankeyNetwork.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(
  dat,
  left,
  right,
  rna_umap,
  adt_umap,
  grouping_variable = NULL,
  grouping_values = NULL,
  min_counts = NULL
)
}
\arguments{
\item{dat}{a data frame with all the relevant columns (the colData object from a SingleCellExperiment or MultiAssayExperiment)}

\item{left}{annotation in the left of the diagram}

\item{right}{annotation in the right of the diagram}

\item{rna_umap}{first dimension reduction data.frame}

\item{adt_umap}{second dimension reduction data.frame}

\item{grouping_variable}{grouping variable found in `dat`}

\item{grouping_values}{values in `grouping_variable` that need to be filtered}

\item{min_counts}{minimum number of cells in a link for this to be displayed}
}
\value{
Returns a list with the different elements for the stankey network.
}
\description{
A fct function to prepare the data for the stankey network
}
